Create Contact.java
This is a java Contact class through which we can add user's last name ,first name and contact number. I  have provided getter and setter for fields and some validation so that last name and first name is never left blank  and contact number is 11 digits long, has to starts with 0, even though I have used String to store contact number but we are allowed to store only digits.